refactor: ResponseWriter 분리
